backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pdfplumber
sentence-transformers
faiss-cpu
transformers
torch